네번째 commit 일단 데이터베이스 다대일 연관관계 조금 구현해놓음. 리뷰게시판 글쓰기, 삭제, 리뷰리스트 가져오기(3가지 케이스) 구현. 아직미완이라 에러있을 수 있음..

+++ src/main/java/teamFRS/FoodRoadSook/member/MemberService.java
@@ -1,10 +1,7 @@
 package teamFRS.FoodRoadSook.member;
 
-import org.springframework.stereotype.Service;
-import teamFRS.FoodRoadSook.exception.NotFoundException;
 
-import javax.transaction.Transactional;
-import java.util.HashMap;
+import teamFRS.FoodRoadSook.exception.NotFoundException;
 
 
 public interface MemberService {
@@ -22,7 +19,7 @@ public interface MemberService {
 
 
     //마이페이지에서 회원 정보 변경 저장
-    boolean member_update(Long id, MemberDTO memberDTO);
+    boolean member_update(int id, MemberDTO memberDTO);
 
     //회원 정보 탈퇴
     boolean member_delete(String user_id);
